requirements.txt
robotframework-browser >= 19.3.0
robotframework >= 7.2.0
robotframework-retryfailed
robotframework-pabot
robotframework-robocop
robotframework-tidy
robotframework-debug
psutil>=5.9.5
mypy
ruff
pre-commit
invoke